Move ResourcesLayout to helpers
